chore: fix clang-17 compiler (#9666) 8a1f3286b9 * chore: fix clang-17 compiler
* chore: adding missing file

* fix: rive_native builds

* chore: remove no runtime linking on linux

* chore: more fixes

* chore: removing rive_common

* chore: use build_rive.sh for ios

* chore: use rive_build.sh for recorder

* chore: fix fill missing version

* chore: add rive_build.sh to path

* chore: add rive_build.sh to pr_ios_tests.yaml

* chore: add rive_build to the recorder tests

* chore: drop rive_flutter tests

* chore: fixing ios tests

* fix misspelled

* chore: cleanup

* chore: use latest zlib

* chore: premake5.lua redirects to premake5_v2

* fix: tvos and ios builds

* fix: unreal build path for miniaudio

+++ include/rive/math/contour_measure.hpp
@@ -27,7 +27,7 @@ public:
         unsigned m_tValue : 30; // Dot30 t value for the end of this segment
         unsigned m_type : 2;    // [private enum]
 
-        float getT() const { return m_tValue * kInvScaleD30; }
+        float getT() const { return (float)m_tValue * kInvScaleD30; }
 
         bool operator<(const Segment& other) const
         {
